Added a card bank
The Card Bank is a component of the Application object, that can be used to Load and store cards. The cards contain only plain old pointers for stuff like textures, that is loaded and deleted by the Content Storage component. The cards are cleaned up in the Card Banks destructor. Also added the necessary types and utilites.