Changed how the player's cards are drawn.
The cards are now drawn horizontally close to the bottom of the screen. The Character is drawn right above the cards. The cards in the player's hand are drawn into a row of the size of 8 cards heads pointing upwards. placed right next to each other horizontally.

Made the card position offset computation happen only once.

Moved the computation out from the for loop for the cards in the player's hand.