Mostly line length changes.
Also introduced an IdFul class that can be used as a parent class and provides a unique ID across all objects that inherit from it. Not thread safe, naive implementation.
Also fixed the CmakeLists.txt and added the Utils lib as a separate dll. Cmake now generate a compile_commands.json file, that can be used with VSCode for include paths.