Windows friendlification
Created an external folder, that is so far not part of the repo, because I want to make it properly, i.e. submodules where applicable and some proper way to bundle prebuilt dependencies.

For some reason the Graphics lib had to be linked to Utils to compile otherwise I've had an unresolved external symbol problem in the IdFul constructor called from Positionable. I also had to un-inline the constructor of IdFul to not have an unresolved external symbol on the NextId member. I don't understand why did I have to make those changes.

I also couldn't find a sympathic way to look for different libs in different configurations.

These things might or might not get fixed in the future.