[Submodule Externals] Having the dependencies as submodules.
[Submodule Externals] Using SDL_ttf CMake package.
[Submodule Externals] Added SDL_ttf as submodule.
[Submodule Externals] Using SDL2Image CMake package.
[Submodule Externals] Added SDL2Image as submodule
[Submodule Externals] Using SDL2 CMake package.
[Submodule Externals] Added SDL2 as a submodule.
[Submodule Externals] JsonCpp used as a cmake package.
Changed the add_subdirectory call to find_package and switched to using the dll version of jsoncpp. Also set BangGame to be the VS startup project and of course added the post build step to copy jsoncpp.dll next to the executable. Also added the vendor folder (where the installed versions of the dependencies are expected) to the .gitignore file.
[Submodule Externals] JSONCPP seems to work
Plus a bit of cleanup of logs.
[Submodule Externals] Added JSONCPP as a submodule