[Persistent Screen] Started adding Positionable to Card.
Did some changes in the Positionable class, so that the destructor doesn't have to iterate through the children and set their parent to nullptr. Also changed the Positionable::RemoveChild function to us the ID of the Positionable instead of the raw pointer. I still don't like this solution, but the children can't be shared or weak ptrs, because they can be added to a parent in the constructor and shared_from_this can't be called in a constructor.